requirements.txt
gevent>=1.0.2,<2.0.0
gevent-websocket>=0.9.5,<1.0.0
leancloud-sdk>=1.0.9,<=2.0.0
Flask>=0.10.1,<1.0.0
Flask-Sockets>=0.1,<1.0
requests>=2.10.0,<3.0
easytrader
tushare
pandas